Product creation by vendor

+++ resources/views/users/index.blade.php
@@ -7,9 +7,11 @@
         <div class="pull-left">
             <h2>{{base64_decode(request()->route('type'))}} Management</h2>
         </div>
+        @can('user-create')
         <div class="pull-right">
             <a class="btn btn-success" href="{{url('users/create/'. request()->route('type'))}}"> Create New {{base64_decode(request()->route('type'))}}</a>
         </div>
+        @endcan
     </div>
 </div>
 
@@ -27,7 +29,9 @@
    <th>Name</th>
    <th>Email</th>
    <th>Mobile</th>
+   @canany('user-list','user-create','user-edit','user-delete')
    <th width="280px">Action</th>
+   @endcan
  </tr>
  @foreach ($data as $key => $user)
   <tr>
@@ -42,14 +46,20 @@
         @endforeach
       @endif --}}
     </td>
+    @canany('user-list','user-create','user-edit','user-delete')
     <td>
        <a class="btn btn-info" href="{{url('users/show/'. request()->route('type').'/'.$user->id)}}"> <i class="far fa-eye"></i></a>
+       @can('user-edit')
        <a class="btn btn-primary" href="{{url('users/edit/'. request()->route('type').'/'.$user->id)}}"> <i class="far fa-edit"></i></a>
+       @endcan
+       @can('user-delete')
         {!! Form::open(['method' => 'DELETE','url' => url('users/delete/'. request()->route('type').'/'.$user->id),'style'=>'display:inline']) !!}
         {!! Form::button('<i class="far fa-trash-alt"></i>', ['type' => 'submit', 'class' =>
         'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
         {!! Form::close() !!}
+        @endcan
     </td>
+    @endcan
   </tr>
  @endforeach
 </table>
